Moved function to misc
